fix class name
This only works by accident as PHP is kind of case insensitive but the autoloader is not. xxx_BillmateInvoice looks like it should be located in Model/Billmate/Invoice.php but this module loads using getModel('billmateinvoice/billmateinvoice') thus the proper name is xxx_Billmateinvoice. When rewriting/extending this class the autoloader wont find it as it is right now.